Validacion de URL (Da error si se pasa un string que no sea una url).  Agregue validacion de existecia de un recurso en metodos get, patch y delete.

+++ src/controllers/post.js
@@ -10,12 +10,22 @@ module.exports = {
     res.json(posts);
   },
 
+  async getPost(req, res) {
+    const id = req.params.id;
+    const post = await Post.findByPk(id);
+
+    if (post) {
+      res.json({ Post: post });
+    } else {
+      res.json("Error. No existe un recurso con esa id");
+    }
+  },
+
   async postAPost(req, res) {
     const { title, content, category, image, creationDate } = req.body;
 
     if (title && content && category && image && creationDate) {
-
-      const imageValidation = validarImagen(image);
+      const imageValidation = await validarImagen(image);
 
       if (!imageValidation) {
         console.log("Imagen no validada");
@@ -30,33 +40,57 @@ module.exports = {
   },
 
   async updatePost(req, res) {
-    const update = req.body;
-    Post.update(update, {
-      where: { id: req.params.id },
-    })
-      .then(() => {
-        res.json("El post ha sido actualizado correctamente");
+    const post = await Post.findByPk(req.params.id);
+    if (!post) {
+      res.json("Error. No existe un recurso con esa id.");
+    } else {
+      const update = req.body;
+      Post.update(update, {
+        where: { id: req.params.id },
       })
-      .catch((err) => {
-        console.log(err);
-        res.json({ Error: err });
-      });
+        .then(() => {
+          res.json("El post ha sido actualizado correctamente");
+        })
+        .catch((err) => {
+          console.log(err);
+          res.json({ Error: err });
+        });
+    }
   },
 
   async deletePost(req, res) {
-    await Post.destroy({
-      where: { id: req.params.id },
-    });
+    const post = await Post.findByPk(req.params.id);
+    if (!post) {
+      res.json("Error. No existe un recurso con ese id.");
+    } else {
+      await Post.destroy({
+        where: { id: req.params.id },
+      });
 
-    res.json({ Succes: "Post has been deleted" });
+      res.json({ Succes: "Post has been deleted" });
+    }
   },
 };
 
-function validarImagen(url) {
+async function validarImagen(url) {
   const regExp = /\.(jpg|png|)$/i;
   if (!regExp.test(url)) {
     return false;
   } else {
-    return true;
+    var http = require("https");
+    var valid = await http
+      .get(url, (res) => {
+        valid = true;
+      })
+      .on("error", (err) => {
+        console.log("Error: " + err.message);
+        valid = false;
+      });
+
+    if (valid) {
+      return true;
+    } else {
+      return false;
+    }
   }
 }
